Add support for command-line arguments
The program now accepts a command-line argument to specify the input CSV file path. If no argument is provided, a default path is used. This should make the program more flexible and easier to use in a variety of contexts.

Additionally, some minor changes were made to error handling to ensure that all errors are propagated up to the main function and printed to the console.